requirements.txt
Django==2.2
django-dotenv==1.4.2
djangorestframework==3.12.2
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1